refactor: extract shared MCP server scaffolding into mcp-common
The context-engineering and uniqueness MCP servers carried near-identical
copies of the transport wiring (stdio / streamable-http / sse beans), the
server-build logic, the tool-registration loop, and the command-line
transport-mode bootstrap. Pull all of it into a new top-level mcp-common
module:

- AbstractMcpConfiguration holds every transport @Bean and the server build;
  concrete servers supply only serverName() and tools().
- McpTool is the shared tool SPI (ContextTool now extends it; UniquenessTool
  implements it).
- TransportConfigurer holds the --transport.mode parsing and runtime priming
  the two Main classes used to duplicate.

Both servers' Main classes shrink to a trivial bootstrap, and each
McpConfiguration to its name plus tool list. The duplicated Main unit tests
are replaced by a single TransportConfigurerTest; the existing
McpConfiguration unit tests and the SSE / streamable-HTTP integration tests
exercise the inherited beans and still pass.
