+++ code/context/pom.xml
@@ -66,6 +66,11 @@
 
 	<dependencies>
 		<dependency>
+			<groupId>io.github.adamw7</groupId>
+			<artifactId>tools.mcp-common</artifactId>
+			<version>${revision}</version>
+		</dependency>
+		<dependency>
 			<groupId>io.modelcontextprotocol.sdk</groupId>
 			<artifactId>mcp-json-jackson2</artifactId>
 		</dependency>
