+++ code/context/src/main/java/io/github/adamw7/context/mcp/ContextTool.java
@@ -1,19 +1,11 @@
 package io.github.adamw7.context.mcp;
 
-import java.util.Map;
-import java.util.function.Function;
-
-import io.modelcontextprotocol.spec.McpSchema.CallToolResult;
-import io.modelcontextprotocol.spec.McpSchema.Tool;
+import io.github.adamw7.tools.mcp.McpTool;
 
 /**
- * A tool the context-engineering MCP server exposes. Each tool carries its own
- * {@link Tool} definition (name and input schema) and maps a call's arguments to
- * a {@link CallToolResult}. Depending on this abstraction lets
- * {@link McpConfiguration} register any number of tools without knowing their
- * concrete types.
+ * A tool the context-engineering MCP server exposes. It is just an
+ * {@link McpTool}; the dedicated type keeps the context tools grouped under one
+ * name and lets {@link McpConfiguration} talk about them in domain terms.
  */
-public interface ContextTool extends Function<Map<String, Object>, CallToolResult> {
-
-	Tool getToolDefinition();
+public interface ContextTool extends McpTool {
 }
