+++ code/context/src/main/java/io/github/adamw7/context/mcp/Main.java
@@ -3,6 +3,8 @@ package io.github.adamw7.context.mcp;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+import io.github.adamw7.tools.mcp.TransportConfigurer;
+
 /**
  * Entry point of the context-engineering MCP server. It selects the transport
  * from {@code --transport.mode} (defaulting to stdio) before starting Spring so
@@ -13,26 +15,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 public class Main {
 
 	public static void main(String[] args) {
-		configureRuntime(args);
+		TransportConfigurer.configure(args);
 		SpringApplication.run(Main.class, args);
-	}
-
-	static void configureRuntime(String[] args) {
-		String transportMode = resolveTransportMode(args);
-		System.setProperty("transport.mode", transportMode);
-		if ("stdio".equals(transportMode)) {
-			System.setProperty("spring.main.web-application-type", "none");
-		}
-		System.setProperty("spring.main.banner-mode", "off");
-	}
-
-	static String resolveTransportMode(String[] args) {
-		String prefix = "--transport.mode=";
-		for (String arg : args) {
-			if (arg.startsWith(prefix)) {
-				return arg.substring(prefix.length());
-			}
-		}
-		return "stdio";
 	}
 }
